+++ shared/GR_PlayerInvite.lua
@@ -37,6 +37,7 @@ end
 function invite:StartInvite(pName, class, useInviteMsg, useWhisperMsg, useGreetingMsg, isManual, sendInvite)
     if not pName then return end
 
+    --pName = 'Monkstrife' --! Remove this line
     local fName = pName:find('-') and pName or pName..'-'..GetRealmName() -- Add realm name if not present
     local name = pName:gsub('*-', '') -- Remove realm name if present
     local cName = class and ns.code:cPlayerName(name, class) or name
@@ -78,9 +79,9 @@ function invite:StartInvite(pName, class, useInviteMsg, useWhisperMsg, useGreeti
     invite:RegisterInvite(pName, class, useWhisperMsg, useGreetingMsg, isManual)
 
     if invFormat ~= 2 and invFormat ~= 4 and useInviteMsg and msgInvite then
-        if isManual then self:SendMessage(pName, name, msgInvite)
+        if isManual or not ns.core.obeyBlockInvites then self:SendMessage(pName, name, msgInvite)
         else
-            C_Timer.After(.5, function()
+            C_Timer.After(1, function()
                 if invite.tblSent[strlower(fName)] then
                     self:SendMessage(pName, name, msgInvite)
                     invite.tblSent[strlower(fName)].sentAt = GetServerTime()